+++ src/components/atoms/CardHero.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import LazyLoad from 'react-lazy-load';
 
-const CardHero = ({ className = '', loadImageViaCss = false, imageSrc, alt, ...htmlAttributes }) => {
+const CardHero = ({ className = '', loadImageViaCss = false, imageSrc, alt, children, ...htmlAttributes }) => {
     if (loadImageViaCss) {
         return (
             <div
@@ -16,7 +17,11 @@ const CardHero = ({ className = '', loadImageViaCss = false, imageSrc, alt, ...h
 
     return (
         <div className={`vdp-card__hero ${className}`}>
-            <img className="vdp-card__hero__image" src={imageSrc} {...htmlAttributes} alt={alt} />
+            {/* height same as $card-hero-height */}
+            <LazyLoad height="128px">
+                <img className="vdp-card__hero__image" src={imageSrc} {...htmlAttributes} alt={alt} />
+            </LazyLoad>
+            {children}
         </div>
     );
 };
@@ -28,7 +33,8 @@ CardHero.propTypes = {
     /** Hero image url */
     imageSrc: PropTypes.string.isRequired,
     /** Defines whether the hero image is loaded  via an image tag or a backgroundImage style property via css */
-    loadImageViaCss: PropTypes.bool
+    loadImageViaCss: PropTypes.bool,
+    children: PropTypes.node
 };
 
 export default CardHero;